+++ gemini.py
@@ -7,7 +7,9 @@ load_dotenv()
 api_key = os.environ["GEMINI_API_KEY"]
 
 # Initialize the model
-model = Gemini(gemini_api_key=api_key)
+model = Gemini(
+    gemini_api_key="AIzaSyCUMFvG3TwwuZIniSpvDeP2AW5USyJUgJM"
+)
 
 # Establish the prompt and image
 task = "What is your name"
