+++ swarms/memory/chroma_db.py
@@ -62,7 +62,7 @@ class ChromaDB:
         top_results_num: int = 3,
         limit_tokens: Optional[int] = 1000,
         *args,
-        **kwargs
+        **kwargs,
     ):
         self.metric = metric
         self.RESULTS_STORE_NAME = RESULTS_STORE_NAME
@@ -93,7 +93,9 @@ class ChromaDB:
             embedding_function=embedding_function,
         )
 
-    def add(self, task: Dict, result: str, result_id: str, *args, **kwargs):
+    def add(
+        self, task: Dict, result: str, result_id: str, *args, **kwargs
+    ):
         """Adds a result to the ChromaDB collection
 
         Args:
@@ -140,19 +142,14 @@ class ChromaDB:
                         "result": result,
                     },
                     *args,
-                    **kwargs
+                    **kwargs,
                 )
         except Exception as error:
             print(
                 colored(f"Error adding to ChromaDB: {error}", "red")
             )
 
-    def query(
-        self,
-        query: str,
-        *args,
-        **kwargs
-    ) -> List[dict]:
+    def query(self, query: str, *args, **kwargs) -> List[dict]:
         """Queries the ChromaDB collection with a query for the top results
 
         Args:
@@ -171,7 +168,7 @@ class ChromaDB:
                 n_results=min(self.top_results_num, count),
                 include=["metadatas"],
                 *args,
-                **kwargs
+                **kwargs,
             )
             out = [item["task"] for item in results["metadatas"][0]]
             out = limit_tokens_from_string(
